% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkggeocoder2-package.R
\docType{data}
\name{commaddr}
\alias{commaddr}
\title{Address data of community center addresses}
\format{
A tibble with 7976 rows and 6 columns:
\describe{
\item{osm_id}{ID of the OSM object linked to the address}
\item{addr.street}{Street name of the address}
\item{addr.housenumber}{House number of the address}
\item{addr.postcode}{Zip code of the address (PLZ)}
\item{addr.city}{Place name of the address (Gemeinde)}
\item{address}{Address string glued together from the other columns}
}
}
\source{
https://www.openstreetmap.org/, https://overpass-turbo.eu/
}
\usage{
commaddr
}
\description{
A dataset containing the addresses of community centers in
Germany. The dataset was extracted from OpenStreetMap on 2022-05-12 19:25.
}
\keyword{datasets}
