% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.GeocodingResults}
\alias{print.GeocodingResults}
\title{Print geocoding results}
\usage{
\method{print}{GeocodingResults}(x, n = 10, ...)
}
\arguments{
\item{x}{Object of class \code{GeocodingResults}}

\item{n}{Maximum number of rows to display. Defaults to 10.}

\item{...}{Further arguments passed on to
\code{\link[base:print.data.frame]{base::print.data.frame()}}}
}
\description{
Print geocoding results
}
