% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.GeocodingResults}
\alias{summary.GeocodingResults}
\title{Get a summary of geocoding results}
\usage{
\method{summary}{GeocodingResults}(object, ...)
}
\arguments{
\item{object}{Object of class \code{GeocodingResults}}

\item{...}{Ignored.}
}
\description{
Get a summary of geocoding results
}
