% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{bkg_export_geocodes}
\alias{bkg_export_geocodes}
\title{Export geocoding results}
\usage{
bkg_export_geocodes(.data, file, overwrite = TRUE, ...)
}
\arguments{
\item{.data}{\code{[GeocodingResults]}

Output of \code{\link[bkggeocoder2]{bkg_geocode_offline}} that should be
exported.}

\item{file}{\code{[character]}

Path to the output file. The file type is guessed based on the file extension.
If the file extension is \code{csv}, the data is exported using
\code{\link[readr]{write_delim}}. If it is \code{xlsx}, the data is exported
using \code{\link[openxlsx]{writeData}}. If is is anything else, the file type
is guessed by \code{\link[sf]{st_write}} and must be supported by
\code{\link[sf]{st_drivers}}.}

\item{overwrite}{\code{[logical]}

Whether to overwrite \code{file}, if it already exists. Defaults to \code{TRUE}.}

\item{...}{Further arguments passed to \code{\link[readr]{write_delim}},
\code{\link[openxlsx]{writeData}} or \code{\link[sf]{st_write}}.}
}
\value{
\code{file}, invisibly.
}
\description{
Export the output of \code{\link[bkggeocoder2]{bkg_geocode_offline}}.
}
