% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.GeocodingResults}
\alias{plot.GeocodingResults}
\title{Plot geocoding score distribution}
\usage{
\method{plot}{GeocodingResults}(x, ...)
}
\arguments{
\item{x}{Object of class \code{GeocodingResults}}

\item{...}{Further arguments passed on to \code{\link[graphics]{hist}}}
}
\description{
Plot geocoding score distribution
}
