% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spt_create_inspire_ids}
\alias{spt_create_inspire_ids}
\title{Create 1km and 100m INSPIRE IDs}
\usage{
spt_create_inspire_ids(
  data,
  type = c("1km", "100m"),
  column_name = "Gitter_ID_",
  combine = FALSE
)
}
\arguments{
\item{data}{Object of class \code{sf} containing point geometries}

\item{type}{Character string for the requested ID type}

\item{column_name}{Output column name prefix. Defaults to "Gitter_ID_\{type\}".}

\item{combine}{Whether to combine the input data with the output values.}
}
\value{
tibble
}
\description{
Create 1 km x 1 km and 100m x 100m INSPIRE IDs from coordinates
}
